+++ tests/readwrite.py
@@ -11,6 +11,9 @@ class FakeGuild:
     def __init__(self):
         self.id = random.randint(1, 1000)
 
+    def __repr__(self) -> str:
+        return str(self.id)
+
 
 class FakeBot:
     def __init__(self):
@@ -22,22 +25,23 @@ class FakeBot:
 class ReadWrite(unittest.TestCase):
     def setUp(self):
         self.bot = FakeBot()
+        self.reader = guildreader.Reader("test", self.bot.guilds)
 
     def step1(self):
-        guildreader.create_file(self.bot, "test", "Yes", wipe=True)
+        self.reader.create_file("test", "Yes", wipe=True)
 
     def step2(self):
         for guild in self.bot.guilds:
-            data = guildreader.read_file(guild.id, "test")
+            data = self.reader.read_file(guild.id, "test")
             assert data == "Yes"
 
     def step3(self):
         for guild in self.bot.guilds:
-            guildreader.write_file(guild.id, "test", "rewrite")
+            self.reader.write_file(guild.id, "test", "rewrite")
 
     def step4(self):
         for guild in self.bot.guilds:
-            data = guildreader.read_file(guild.id, "test")
+            data = self.reader.read_file(guild.id, "test")
             assert data == "rewrite"
 
     def _steps(self):
